Move authentication to its own file
